Fix mkbundle on Mac OS X
On Mac OS X files got trailing zeros which somehow seam to work
quite nice except for the javascript parser that did not like them.
This adds a awk that filters out empty lines from od and also
fixes some of the sed substitutions that seamed broken.

Tested manually on Linux and it seams to output correct C code.