upnp: Rework video playback from UPNP
Instead of storing the target URL in the model for video items we
now store an URL pointing to the UPNP object. Upon browsing that
we query the UPNP server once again. This allows us to do an
additional query for subtitles on minidlna servers. In the future
we might also come up with other ways of searching for subs.