Redesign Showtime's buffer management for streaming.
Showtime now use a fixed amount of memory for buffering demultiplexed
media packets. On top of this there is also a realtime delay parameter
that can be used to limit the buffer depth to a fixed time. The latter
is useful for RTMP streaming where the server only allows us to
per-buffer a certain amount of time and we don't want to block on
server socket reads for performance and UX reasons.

The max buffer memory is set by Showtime based on the type of media
being played:

 10 MB: Used for video streaming
 1 MB: Used for audio playback
 0: Used for in-memory playback (SID, NSF, SPC, etc)

The audio / subtitle track filtering has been moved to after the
queue which should cut the perceived delay between user input
and track switch to almost nothing. Without this change and
with the new buffering the delay would have otherwise been ~10s.

Should fix ticket #492 #491