Don't wait for the video playback controller to terminate.
We don't need to as we have refcounted media_pipe_t anyway.
This removes a point where the UI would freeze waiting for the
video playback to terminate. Even worse, if the video playback
would have stalled for an extended period of time (internet badness)
it would appear to lock the UI completely

Fixes #790