Fix play/pause issues
When pausing audio and then playing a different track the new track would
not start (but rather be in paused state). However the UI dispalyed that
the track should be playing so confusion was total.

Now, just play the track instead (unpause it)

Also fix a crash caused by inconsistency between the pulseaudio driver
and the common audio code (caused by discarding MB_CTRL_UNBLOCK in flush)